+++ wolframclient/language/expression.py
@@ -99,14 +99,14 @@ class WLFunction(WLExpressionMeta):
 class WLSymbolFactory(WLSymbol):
     """Provide a convenient way to build objects representing arbitrary Wolfram Language expressions through the use of attributes.
 
-    This class is conveniently instanciated at startup as: :class:`~wolframclient.language.wl`, :class:`~wolframclient.language.Global` 
-    and :class:`~wolframclient.language.System`. It should be instanciated only to represent many symbols belonging to the same specific
+    This class is conveniently instantiated at startup as: :class:`~wolframclient.language.wl`, :class:`~wolframclient.language.Global` 
+    and :class:`~wolframclient.language.System`. It should be instantiated only to represent many symbols belonging to the same specific
     context.
 
     Example::
 
-        >>> developer = WLSymbolFactory('Developer')
-        >>> developer.PackedArrayQ
+        >>> dev = WLSymbolFactory('Developer')
+        >>> dev.PackedArrayQ
         Developer`PackedArrayQ
     
     Alternative::
@@ -139,3 +139,22 @@ class WLInputExpression(WLExpressionMeta):
 
     def __str__(self):
         return '(%s)' % self.input
+
+
+# Sphinx seems to bug on this one, and picks an outdated the docstring when declared in __init__.
+wlexpr = WLInputExpression
+""" Represent Wolfram Language expressions with input form strings.
+
+Convenient alias for :class:`~wolframclient.language.expression.WLInputExpression`.
+
+Represent an expression::
+
+    >>> wlexpr('Select[EvenQ, Range[10]]')
+    (Select[EvenQ, Range[10]])
+
+Represent a pure function that squares an input argument::
+
+    >>> wlexpr('# ^ 2 &' )
+    (# ^ 2 &)
+
+"""