+++ wolframclient/serializers/utils.py
@@ -33,7 +33,7 @@ def py_encode_text(s):
     yield b'"'
 
 
-def py_encode_decimal(number, prec = decimal.getcontext().prec):
+def py_encode_decimal(number, prec=decimal.getcontext().prec):
     return '{0:f}``{1:d}'.format(number, prec).encode('utf-8')