+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# Version 1.0.0.dev4
+# Version 1.0.0
 - Inconsistent class name `WolframCloudSessionFuture` replaced with `WolframCloudFutureSession`.
+- Rework completely `Dispatch` class. Add a `Dispatch` instance called `wolfram_encoder` representing the multi dispatch function that encodes Python object.
+- Add an entry point allowing new encoders to be registered as separated libraries (plugins).
+- Add support for pandas classes: `Series` and `DataFrame`.
+- Extend support for NumPy numbers.
 
 # Version 1.0.0.dev3
 - Add `CHANGELOG.md` to the repository to help users track the changes.